requirements/base.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.1
cffi==1.15.0
cfgv==3.3.1
cryptography==36.0.0
distlib==0.3.3
Django==3.2.9
django-filter==21.1
django-guardian==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
filelock==3.4.0
flake8==4.0.1
identify==2.4.0
importlib-metadata==4.8.2
iniconfig==1.1.1
Markdown==3.3.6
mccabe==0.6.1
nodeenv==1.6.0
packaging==21.3
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.15.0
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==3.0.6
pytest==6.2.5
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.10.0
zipp==3.6.0
